Fine tune --version, --help and manual page
Need to do some custom formatting in make-man.sh with help2man 1.47.15
to make the usage (now in SYNOPSIS chapter) look good, to disable
hyphenation of URLs and to have some line break.

Debug information is not stored under /tmp anymore by default so fix
that in documentation. Remove the needless link
https://www.gnu.org/software/gethelp.html in there, too, but add
author in the output of --version and thus in manual page, as well.