Improve debug file handling
Show debug filename only, if it is not given by user.

Do not try to remove an empty debug file, if it is not regular. (Debug
file could be e.g. terminal whose name is given by tty command in that
terminal.)

Change the owner of the debug file, if needed, in order to make it
accessible without sudo even when the command is run as superuser.