Improves behavior of --info
Uses new entire_pipe() function to run the pipe of show_info() in order
to avoid broken pipe error if the pager used quits early. This is
supposed to work even if PIPE trap has been set null string in the
shell calling this program.

When --debug is used with --info, operate whole show_info() by debug(),
not just get_info(). This has the side effect that pager runs non-
interactively, but on the other hand it shows more debug information.

Put setting of pipefail option in a subshell so no need to worry about
restoring the setting even in case of an error.