Handle missing Auto-Installed field and do some file validation
Talking about apt's extended_states file that is read in
get_manual_releases(). Normally field Auto-Installed has value 1 for
each package listed in the file, but value 0 marks manually installed
package. And if the field is missing, the package is manually
installed. Packages not mentioned in the file are manually installed,
too.

Stop searching for Auto-Installed field for a package after an empty
line and exit the function with status 5 if the value is invalid.

Exit the function with status 6 if no package is detected in the file.