Add --no-booted option to support running in containers
When running in e.g. a Docker container, the booted kernel is of the enclosing
system and probably not installed in the container itself. Thus we do not
need to check for booted kernel there. A container might not satisfy all
dependencies of this software so try to take that into account.
Fixes LP: #1949504